Add a package listing detail view
Implement a package listing detail view which takes care of commonly
used functionality such as permission checks and tab context for the
package detail views.
